requirements.txt
alembic==1.13.2
click==8.1.7
Flask==1.1.2
Flask-Migrate==4.0.7
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.0.1
greenlet==3.0.3
itsdangerous==1.1.0
Jinja2==2.11.3
Mako==1.3.5
MarkupSafe==1.1.1
numpy==2.1.0
pandas==2.2.2
psycopg2-binary==2.9.9
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
six==1.16.0
SQLAlchemy==1.4.47
typing_extensions==4.12.2
tzdata==2024.1
Werkzeug==0.16.1
WTForms==3.1.2
